+++ example/index.ts
@@ -1,11 +1,40 @@
 import PocketBase from 'pocketbase';
-import { CollectionRecords } from './pocketbase-types';
-import { TypedPocketBase, neq, sort, fields } from '../src';
+import { Schema } from './Database.js';
+import { TypedPocketBase, fields, expand, eq, asc } from '../src/index.js';
 
-const db: TypedPocketBase<CollectionRecords> = new PocketBase('http://localhost:8090');
+const db = new PocketBase('http://localhost:8090') as TypedPocketBase<Schema>;
+await db.admins.authWithPassword('test@test.com', 'secretpassword');
 
-const { items } = await db.collection('posts').getList(1, 10, {
-	fields: fields('content', 'created', 'id'),
-	sort: sort('-date'),
-	filter: neq('content', '')
-});
+{
+	const posts = await db.collection('posts').getFullList({
+		fields: fields('id', 'title', 'slug', 'content'),
+		sort: asc('date'),
+		filter: eq('published', true)
+	});
+
+	console.log(posts);
+}
+
+{
+	const posts = await db.collection('posts').getFullList({
+		sort: asc('date'),
+		filter: eq('owner.email', 'user@test.com'),
+		expand: expand({
+			owner: true
+		})
+	});
+
+	console.log(posts[0].expand.owner);
+}
+
+{
+	const post = await db
+		.collection('posts')
+		.getFirstListItem(eq('owner.email', 'user@test.com'), {
+			expand: expand({
+				owner: true
+			})
+		});
+
+	console.log(post);
+}