+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # typed-pocketbase
 
+## 0.0.3
+
+### Patch Changes
+
+-   dd52f65: rework codegen cli and support all field types and modifiers
+-   4201a28: export TypedRecord
+-   5b60b0c: fix entrypoints and file thumbs type when none are specified
+-   376dc88: Add expand and nested filter support
+
+## 0.0.3-next.3
+
+### Patch Changes
+
+-   rework codegen cli and support all field types and modifiers
+
+## 0.0.3-next.2
+
+### Patch Changes
+
+-   export TypedRecord
+
+## 0.0.3-next.1
+
+### Patch Changes
+
+-   fix entrypoints and file thumbs type when none are specified
+
+## 0.0.3-next.0
+
+### Patch Changes
+
+-   Add expand and nested filter support
+
 ## 0.0.2
 
 ### Patch Changes